Update 2.1 additions and fix
-fixed crash by using 32bit long
-changed 32bit to 64bit long in Injector and NativeAPI
-added setDebug to Injector
-changed build.add<long> to long long
-changed from AutomaticID* to AutomaticID in Level.h this just work
-added destructor to DataOffset and fixed free method(called destructor)

+++ src/native/main/core/engine/module/hook_java.cpp
@@ -167,7 +167,7 @@ const ArgsBufferBuilder HookJava::getParameters(JNIEnv* env, void* self, std::ve
         }else if(type == "float"){
             builder.add<float>(NativeAPI::getFloatHookParameter(env, object));
         }else if(type=="long"){
-            builder.add<long>(NativeAPI::getLongHookParameter(env,object));
+            builder.add<long long>(NativeAPI::getLongHookParameter(env,object));
         }else{
             builder.add<void*>(NativeAPI::getPointerHookParameter(env, object));
         }
